Remove Post endpoints for an editon and version
These endpoints have been removed as the creation of an edition and a
version document occurs off the back of an instance state changing to
`edition-confirmed`
