+++ dimension/dimension_test.go
@@ -7,12 +7,13 @@ import (
 	"net/http/httptest"
 	"testing"
 
+	"strings"
+
 	errs "github.com/ONSdigital/dp-dataset-api/apierrors"
 	"github.com/ONSdigital/dp-dataset-api/dimension"
 	"github.com/ONSdigital/dp-dataset-api/models"
 	"github.com/ONSdigital/dp-dataset-api/store/datastoretest"
 	. "github.com/smartystreets/goconvey/convey"
-	"strings"
 )
 
 const secretKey = "coffee"
@@ -89,7 +90,7 @@ func TestAddDimensionToInstanceReturnsOk(t *testing.T) {
 	t.Parallel()
 	Convey("Add a dimension to an instance returns ok", t, func() {
 		w := httptest.NewRecorder()
-        json := strings.NewReader(`{"value":"24", "code_list":"123-456", "dimension_id": "test"}`)
+		json := strings.NewReader(`{"value":"24", "code_list":"123-456", "dimension_id": "test"}`)
 		r := createRequestWithToken("POST", "http://localhost:21800/instances/123/dimensions", json)
 		mockedDataStore := &storetest.StorerMock{
 			AddDimensionToInstanceFunc: func(event *models.CachedDimensionOption) error {
@@ -264,95 +265,5 @@ func TestGetUniqueDimensionValuesReturnsInternalError(t *testing.T) {
 
 		So(w.Code, ShouldEqual, http.StatusInternalServerError)
 		So(len(mockedDataStore.GetUniqueDimensionValuesCalls()), ShouldEqual, 1)
-	})
-}
-
-func TestGetDimensionsReturnsOk(t *testing.T) {
-	t.Parallel()
-	Convey("When the request contain valid ids return dimension information", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123/editions/2017/versions/1/dimensions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			GetDimensionsFunc: func(datasetID, editionID, versionID string) (*models.DatasetDimensionResults, error) {
-				return &models.DatasetDimensionResults{}, nil
-			},
-		}
-		dimension := &dimension.Store{mockedDataStore}
-		dimension.GetDimensions(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-}
-
-func TestGetDimensionsReturnsErrors(t *testing.T) {
-	t.Parallel()
-
-	Convey("When the request contain invalid ids return not found error", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123/editions/2017/versions/1/dimensions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			GetDimensionsFunc: func(datasetID, editionID, versionID string) (*models.DatasetDimensionResults, error) {
-				return nil, errs.DatasetNotFound
-			},
-		}
-		dimension := &dimension.Store{mockedDataStore}
-		dimension.GetDimensions(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-	})
-	Convey("When the api cannot connect to datastore to get dimension resource return an internal server error", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123/editions/2017/versions/1/dimensions", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			GetDimensionsFunc: func(datasetID, editionID, versionID string) (*models.DatasetDimensionResults, error) {
-				return nil, internalError
-			},
-		}
-		dimension := &dimension.Store{mockedDataStore}
-		dimension.GetDimensions(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
-	})
-}
-
-func TestGetDimensionOptionsReturnsOk(t *testing.T) {
-	t.Parallel()
-	Convey("", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123/editions/2017/versions/1/dimensions/age/options", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			GetDimensionOptionsFunc: func(datasetID, editionID, versionID, dimensions string) (*models.DimensionOptionResults, error) {
-				return &models.DimensionOptionResults{}, nil
-			},
-		}
-		dimension := &dimension.Store{mockedDataStore}
-		dimension.GetDimensionOptions(w, r)
-		So(w.Code, ShouldEqual, http.StatusOK)
-	})
-}
-
-func TestGetDimensionOptionsReturnsErrors(t *testing.T) {
-	t.Parallel()
-
-	Convey("", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123/editions/2017/versions/1/dimensions/age/options", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			GetDimensionOptionsFunc: func(datasetID, editionID, versionID, dimensions string) (*models.DimensionOptionResults, error) {
-				return nil, errs.DatasetNotFound
-			},
-		}
-		dimension := &dimension.Store{mockedDataStore}
-		dimension.GetDimensionOptions(w, r)
-		So(w.Code, ShouldEqual, http.StatusNotFound)
-	})
-	Convey("", t, func() {
-		r := httptest.NewRequest("GET", "http://localhost:22000/datasets/123/editions/2017/versions/1/dimensions/age/options", nil)
-		w := httptest.NewRecorder()
-		mockedDataStore := &storetest.StorerMock{
-			GetDimensionOptionsFunc: func(datasetID, editionID, versionID, dimensions string) (*models.DimensionOptionResults, error) {
-				return nil, internalError
-			},
-		}
-		dimension := &dimension.Store{mockedDataStore}
-		dimension.GetDimensionOptions(w, r)
-		So(w.Code, ShouldEqual, http.StatusInternalServerError)
 	})
 }
